requirements.txt
numpy
torch
torchvision
blobfile
tqdm
pyYaml
pillow